This duplicate application was picked up as an observation
The client would like to see the rules like this. All the text in 1
block

With the submitted column on the above form, can you have a Submitted
Date+Time on the Application form?

I am thinking we have that support message defined in the Program
Definition, so that we can change it by program?